+++ packages/mint-components/src/components/sqm-portal-login/sqm-portal-login.tsx
@@ -78,8 +78,8 @@ export class PortalLogin {
   @Prop() pageLabel: string = "Sign in to your account";
 
   /**
-   * @uiName Network error message
    * Displayed when the login fails due to a network error. The participant can try refreshing the page.
+   * @uiName Network error message
    */
   @Prop() networkErrorMessage: string =
     "An error occurred while logging you in. Please refresh the page and try again.";